requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
isort==5.10.1
jedi==0.18.2
multidict==6.0.2
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.2
platformdirs==2.5.4
python-dotenv==0.21.0
requests==2.28.1
smmap==5.0.0
urllib3==1.26.12
yarl==1.8.1
